refactor: move encrypt/decrypt deoxys fns to encryption.rs
